requirements.txt
kafka-python
requests
pyspark
cassandra-driver
statsmodels
pandas
numpy